Moved sounds to separate cog class
