+++ core.py
@@ -122,6 +122,7 @@ class Core(commands.Cog, name='Core'):
 intents = discord.Intents.all()
 bot = commands.Bot(command_prefix= '.', intents=intents)
 bot.load_extension('stuff')
+bot.load_extension('sounds')
 bot.load_extension('ytplayer')
 # bot.load_extension('timer')
 bot.add_cog(Core(bot))
