+++ stuff.py
@@ -72,67 +72,6 @@ class Stuff(commands.Cog):
 
         self.timezone = pytz.timezone(zone)
         await ctx.send('Time zone set to: ' + str(self.timezone))
-       
-
-    @commands.command(name='sound')
-    async def _sound(self,ctx: commands.Context,soundType):
-        """Arrives to you and plays your favourite sound"""
-
-        if len(self.bot.voice_clients) == 0 and soundType != None:
-            member = ctx.message.author
-            await ctx.send("Playing now Darude-Sandstorm")
-            for channel in member.guild.channels:
-                if channel.type == discord.ChannelType.voice:
-                    for m in channel.voice_states:
-                        if m == member.id:
-                            path = 'audio/'+soundType+'.mp3'
-                            if os.path.isfile(path):
-                                self.voice = await channel.connect()
-                                await asyncio.sleep(0.5)      
-                                self.voice.play(discord.FFmpegPCMAudio(path), after = self.voice.stop())
-                                while(self.voice.is_playing()):
-                                    await asyncio.sleep(1)
-                                await self.voice.disconnect()
-
-    @commands.command(name='soundlist')
-    async def _soundList(self, ctx: commands.Context):
-        """Prints the list of all avaible sounds"""
-
-        embed = discord.Embed(title="Sounds list",description="All of the sounds",color=0x076500)
-
-        sounds = os.listdir('./audio')
-
-        i = 1
-        for sound in sounds:
-            val = sound[:-4]
-            embed.add_field(name=str(i) ,value=val,inline=True)
-            i+=1
-
-        await ctx.send(embed=embed)
-
-    @commands.command(name='soundupload')
-    @commands.has_permissions(manage_guild=True)
-    async def _uploadSound(self, ctx: commands.Context):
-        """Attach a sound to be uploaded with this command"""
-
-        if len(ctx.message.attachments) == 1:
-            file = ctx.message.attachments[0].filename
-            path = "audio/{}".format(file)
-            if file.endswith(".mp3") and not os.path.isfile(path):
-                await ctx.message.attachments[0].save(fp=path)
-
-    @commands.command(name='soundrm')
-    @commands.has_permissions(manage_guild=True)
-    async def _removeSound(self, ctx: commands.Context, sound):
-        """Removes a given sound"""
-
-        try:
-            path = './audio/'+sound+'.mp3' 
-            os.remove(path) 
-            await ctx.send('Removed: '+sound)
-        except Exception:
-            raise
-
     
     #---Listener section---#
     @commands.Cog.listener()
